Implemented Shoe in game.py
Game.py now uses the shoe. Also added code to build a new shoe if the remaining cards is less than 52. Perhaps this logic should compare the remaining cards to a number calculated based on the maximum amount of cards the next hand will be expected to use, or maybe there is a casino standard for this situation.